requirements.txt
torch
transformers
accelerate
bitsandbytes
faiss-cpu
sentence-transformers
pyyaml
rich
psutil
scikit-learn